Create a command line interface for utilities
Database maintenance shouldn't be taken lightly.  I created the shell of
a Node app that will run maintenance functions on the database from the
command line.  It can be executed in two ways:

$ npm run utility
    - or -
$ node utilities/db_utility.js

It doesn't do anything yet!  But, ultimately, it will connect to the
database, and offer several functions with which to interact with the
database.